requirements.txt
pyyaml
click
pytest
python-json-logger
rich